pyproject.toml
# --- PROJECT CONFIGURATION --- #

[build-system]
requires = ["setuptools>=67", "wheel", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

# Metadata (see https://peps.python.org/pep-0621/)
[project]
name = "satellite-consumer"
dynamic = ["version"] # Set automtically using git: https://setuptools-git-versioning.readthedocs.io/en/stable/
description = "Tool for converting raw EUMETSAT NAT files to Zarr"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.12.0"
license = {text = "MIT License"}
authors = [
    { name = "Open Climate Fix Team", email = "info@openclimatefix.org" },
    { name = "Sol Cotton", email = "sol@openclimatefix.org" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "Cartopy == 0.24.1",
    "eumdac == 3.0.0",
    "loguru == 0.7.3",
    "pyproj == 3.7.0",
    "pyresample == 1.32.0",
    "pyyaml == 6.0.2",
    "requests == 2.32.3",
    "satpy[all] == 0.54.0",
    "typer-slim == 0.15.1",
    "xarray == 2025.1.2",
    "zarr == 3.0.2",
]

[dependency-groups]
dev = [
    # Testing
    "unittest-xml-reporting",
    # Linting and type checking
    "mypy",
    "ruff >= 0.9.2",
    # LSP Support
    "python-lsp-server",
    "pylsp-mypy",
    "python-lsp-ruff",
]

[project.scripts]
# Put entrypoints in here
sat-consumer-cli = "satellite_consumer.cmd.main:cli_entrypoint"

[project.urls]
repository = "https://github.com/openclimatefix/satellite-consumer"

[tool.setuptools]
include-package-data = false

[tool.setuptools-git-versioning]
enabled = true

# --- LINTING AND TYPING CONFIGURATION --- #

# MyPy configuration
# * See https://mypy.readthedocs.io/en/stable/index.html
[tool.mypy]
python_version = "3.12"
dmypy = true
strict = true
warn_unreachable = true
warn_return_any = true
disallow_untyped_defs = true
plugins = [
    "numpy.typing.mypy_plugin",
]

# Ruff configuration
# * See https://beta.ruff.rs/docs/
[tool.ruff]
line-length = 100
indent-width = 4
exclude = ["__init__.py"]

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "E",   # pycodestyle
    "W",   # whitespace and newlines
    "I",   # isort
    "UP",  # modernize
    "ANN", # flake8 type annotations
    "S",   # flake8 bandit
    "B",   # flake8 bugbear
    "C4",  # flake8 comprehensions
    "COM", # flake8 commas
    "T20", # flake8 print
    "SIM", # flake8 simplify
    "ARG", # flake8 unused arguments
    "DTZ", # flake8 datetimes
    "Q",   # flake8 quotes
    "TCH", # flake8 typecheck
    "D",   # pydocstyle
    "RUF", # ruff-specific rules
]
fixable = ["ALL"]
ignore = [
    "D203", "D213", "D215", "D400", "D401", "D404", "D406",
    "D407", "D408", "D409", "D413",
]

[tool.ruff.lint.per-file-ignores]
"test*" = ["D", "ANN"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 100

# --- DOCUMENTATION CONFIGURATION --- #

[tool.pydoctor]
add-package = ["src/ocf-python-project"]
project-base-dir = "src/ocf-python-project"
docformat = "google"
html-output = "docs"
theme = "classic"
privacy = [
    "HIDDEN:**.test_*",
]
